Chunk data part storage

+++ mch/src/main/java/ca/bkaw/mch/chunk/parts/ChunkDataPart.java
@@ -2,10 +2,19 @@ package ca.bkaw.mch.chunk.parts;
 
 import ca.bkaw.mch.nbt.NbtCompound;
 
+/**
+ * A handler that extracts and merges a part of the nbt data of a chunk.
+ * <p>
+ * Different instances of this class will split chunk data into different parts
+ * that are stored separately. That way only the data part that has changed needs
+ * to be saved again.
+ *
+ * @see ChunkDataParts
+ */
 public abstract class ChunkDataPart {
-    private final int id;
+    private final byte id;
 
-    public ChunkDataPart(int id) {
+    public ChunkDataPart(byte id) {
         this.id = id;
     }
 
@@ -14,7 +23,7 @@ public abstract class ChunkDataPart {
      *
      * @return The id.
      */
-    int getId() {
+    byte getId() {
         return this.id;
     }
